+++ tsdb/engine/tsm1/tsm1.go
@@ -204,7 +204,11 @@ func (e *Engine) PerformMaintenance() {
 		}
 	}
 
-	go e.Compact(true)
+	go func() {
+		if err := e.Compact(true); err != nil {
+			e.logger.Printf("PerformMaintenance: error during compaction: %v", err)
+		}
+	}()
 }
 
 // Format returns the format type of this engine
@@ -320,6 +324,8 @@ func (e *Engine) Close() error {
 	defer e.writeLock.UnlockRange(min, max)
 	e.filesLock.Lock()
 	defer e.filesLock.Unlock()
+
+	e.WAL.Close()
 
 	// ensure all deletes have been processed
 	e.deletesPending.Wait()
@@ -482,7 +488,11 @@ func (e *Engine) Write(pointsByKey map[string]Values, measurementFieldsToSave ma
 	}
 
 	if !e.SkipCompaction && e.shouldCompact() {
-		go e.Compact(false)
+		go func() {
+			if err := e.Compact(false); err != nil {
+				e.logger.Printf("Write: error during compaction: %v", err)
+			}
+		}()
 	}
 
 	return nil